extensions/kd-animation/requirements.txt
#diffusers==0.20.2
#torch 
#torchvision
#torchaudio
#accelerate
clean-fid
#colab-convert
einops
ftfy
ipython
ipywidgets
jsonmerge
#jupyterlab
#jupyter_http_over_ws
kornia
matplotlib
#notebook
numexpr
#omegaconf
#opencv-python
pandas
pytorch_lightning
resize-right
scikit-image
scikit-learn
timm
torchdiffeq
#transformers
#safetensors
#albumentations
more_itertools
devtools
validators
numpngw
open-clip-torch
torchsde
ninja
pydantic
#xformers
imageio[ffmpeg]